commit add

+++ L9FunctionPractice/L9FunctionPractice/L9FunctionPractice/Average3Sums.java
@@ -1,16 +1,16 @@
-// package L9FunctionPractice;
 
 import java.util.Scanner;
 
 public class Average3Sums {
+
     public static void main(String[] args) {
         try (Scanner input = new Scanner(System.in)) {
             System.out.print("Enter the first number: ");
-            int x = input.nextInt();
+            double x = input.nextDouble();
             System.out.print("Enter the second number: ");
-            int y = input.nextInt();
+            double y = input.nextDouble();
             System.out.print("Enter the third number: ");
-            int z = input.nextInt();
+            double z = input.nextDouble();
             double average = (x + y + z) / 3;
             System.out.println("Average is: " + average);
         }
